Make two versions of Location::length(), inclusive and exclusive of end
Add a new Location::length_exclusive() function, to return the length of a
range not including its end point.

Rename Location::length() to length_inclusive(), and use it wherever
Location::length() was used before, so that this whole chenge is a no-op.
Looking at which of these should rather be length_exclusive() is left for
another time.